Fork 0xSplits into our codebase
There were several alternatives to folding this code in, but this
seems like the best method, of just merging these contracts in before we
modify them to suit our needs. The key structures we need to maintain
are:

- split creation logic
- split config hashing and storage
- distribution math

A future commit will rewrite relevant tests for the above from the
splits codebase.