Revert "Implement authorizer check for config updates"
This reverts commit 02888825b20e9b7f4804f604ffdb5b1d2faef19f.

Splits are provided as-is and it is the artist's responsibility to
ensure they are sharing revenue with producers and distributors and
collaborators with the most recent, correct proportions. We will provide
everything needed for artists to do this in the frontend, but we do not
need to authorize any config updates onchain. We just need to prompt a
user to update a config when it falls out of date and provide the tx
data containing the updated split config for them.