Remove artist edition metadata from auction storage
Now that we are minting the artist edition at the creation of an
auction, it is redundant to store it alongside the collector edition
metatdata. This means that the payout address used at the end of the
auction will stored inside the Collector edition metadata, but then not
published as part of the collector edition. This is a bit of a hack, but
given the evm premium on unnecessary data space, it's fine in this context